file injection implemented

+++ gulpfile.js
@@ -1,13 +1,18 @@
 const gulp = require('gulp');
 const nodemon = require('gulp-nodemon');
 const browserSync = require('browser-sync').create();
+const inject = require('gulp-inject');
+const concat = require('gulp-concat');
+const uglify = require('gulp-uglify');
+const es = require('event-stream');
 
 let apiServer = null;
 
 config = {
     js: './src/js/**/*.js',
     css: './src/css/**/*.css',
-    html: './src/**/*.html'
+    html: './src/**/*.html',
+    dist: './dist'
 }
 
 gulp.task('start:api', (cb) => {
@@ -19,7 +24,50 @@ gulp.task('start:api', (cb) => {
     cb();
 });
 
-gulp.task('serve', (cb) => {
+// gulp.task('styles:dev', () => {
+//     return gulp.src([config.css]);
+// });
+
+// gulp.task('scripts:dev', () => {
+//     return gulp.src([config.js]);
+// });
+
+gulp.task('inject:dev', () => {
+    var sources = gulp.src([config.css, config.js], {read: false});
+    return gulp.src('./src/index.html')
+    .pipe(inject(sources, {relative: true}))
+    .pipe(gulp.dest('./src'));
+});
+
+gulp.task('styles:build', () => {
+    return gulp.src([config.css])
+        .pipe(gulp.dest(config.dist));
+});
+
+gulp.task('scripts:build', () => {
+    return gulp.src([config.js])
+        .pipe(concat('gaugeWidjet.js'))
+        .pipe(uglify())
+        .pipe(gulp.dest(config.dist));
+});
+
+gulp.task('inject:build', ['scripts:build', 'styles:build'], () => {
+    var sources = gulp.src([config.dist + '/**/*.css', config.dist + '**/*.js'], {read: false});
+    return gulp.src('./src/index.html')
+        .pipe(inject(sources,{relative: true}))
+        .pipe(gulp.dest(config.dist));
+});
+
+gulp.task('serve:build', ['inject:build'],() => {
+    browserSync.init({
+        server: {
+            port: 3005,
+            baseDir: './dist'
+        }
+    })
+});
+
+gulp.task('serve:dev',['inject:dev'] ,(cb) => {
     browserSync.init({
         server: {
             port: 3005,
@@ -28,14 +76,13 @@ gulp.task('serve', (cb) => {
     });
     cb();
 });
-gulp.task('watch', ['serve'],(cb) => {
+gulp.task('watch', ['serve:dev'], () => {
     gulp.watch('./src/**/*').on('change', browserSync.reload);
-    cb();
 });
 
 process.on('exit', function () {
     // In case the gulp process is closed (e.g. by pressing [CTRL + C]) stop both processes
-    apiServer.kill();
+    apiServer && apiServer.kill();
 });
 
-gulp.task('default', ['watch', 'start:api']);
+gulp.task('default', ['watch']);
